feat: add new card style in blog list
